# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Booking API
  version: 0.0.0
tags: []
paths:
  /event-types:
    get:
      operationId: EventTypes_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
    post:
      operationId: EventTypes_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventTypeRequest'
  /event-types/{id}:
    put:
      operationId: EventTypeById_update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventTypeRequest'
    delete:
      operationId: EventTypeById_delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events:
    get:
      operationId: Events_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      operationId: Events_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
  /events/{id}:
    delete:
      operationId: EventById_delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /slots:
    get:
      operationId: Slots_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
  /slots/generate:
    post:
      operationId: Slots_generate
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateSlotsRequest'
components:
  schemas:
    CreateEventRequest:
      type: object
      required:
        - slotId
        - guestName
        - eventTypeId
        - startTime
        - endTime
        - description
      properties:
        slotId:
          type: string
        guestName:
          type: string
        eventTypeId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        description:
          type: string
    CreateEventTypeRequest:
      type: object
      required:
        - id
        - name
        - description
        - duration
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        duration:
          type: integer
          format: int32
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    Event:
      type: object
      required:
        - id
        - slotId
        - guestName
        - eventTypeId
        - startTime
        - endTime
        - description
      properties:
        id:
          type: string
        slotId:
          type: string
        guestName:
          type: string
        eventTypeId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        description:
          type: string
    EventType:
      type: object
      required:
        - id
        - name
        - description
        - duration
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        duration:
          type: integer
          format: int32
    GenerateSlotsRequest:
      type: object
      required:
        - startDate
        - endDate
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    Guest:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    Owner:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    Slot:
      type: object
      required:
        - id
        - startTime
        - endTime
      properties:
        id:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    UpdateEventTypeRequest:
      type: object
      required:
        - name
        - description
        - duration
      properties:
        name:
          type: string
        description:
          type: string
        duration:
          type: integer
          format: int32
servers:
  - url: /api
    description: API for booking meeting slots
    variables: {}